Move files that contain static data to a seperate directory. Rename device.js to devices.js
