+++ bin/taiko.js
@@ -8,7 +8,7 @@ const { spawnSync } = require('child_process');
 const taiko = require('../lib/taiko');
 const repl = require('../lib/repl');
 const { isTaikoRunner } = require('../lib/util');
-const devices = require('../lib/device').default;
+const devices = require('../lib/data/devices').default;
 let repl_mode = false;
 let plugins = new Map();
 
@@ -100,7 +100,7 @@ if (isTaikoRunner(process.argv[1])) {
         .option('-w, --wait-time <time in ms>', 'runs script with provided delay\n', parseInt)
         .option(
             '--emulate-device <device>',
-            'Allows to simulate device viewport. Visit https://github.com/getgauge/taiko/blob/master/lib/device.js for all the available devices\n',
+            'Allows to simulate device viewport. Visit https://github.com/getgauge/taiko/blob/master/lib/devices.js for all the available devices\n',
             setupEmulateDevice
         )
         .option(
