+++ lib/data/devices.js
