+++ lib/data/networkConditions.js
