+++ lib/inputHandler.js
@@ -18,7 +18,7 @@
  * Few modifications are done on the file.
  */
 
-const keyDefinitions = require('./USKeyboardLayout');
+const keyDefinitions = require('./data/USKeyboardLayout');
 const { assert } = require('./helper');
 let input;
 let _modifiers = 0;
