+++ lib/networkHandler.js
@@ -1,5 +1,5 @@
 const { isFunction, isString, isObject} = require('./helper');
-const networkPresets = require('./networkConditions').default;
+const networkPresets = require('./data/networkConditions').default;
 let requestPromises;
 const defaultErrorReason = 'Failed';
 let interceptors = [];
