+++ lib/taiko.js
@@ -17,7 +17,7 @@ const runtimeHandler = require('./runtimeHandler');
 const browserHandler = require('./browserHandler');
 const emulationHandler = require('./emulationHandler');
 const { handleRelativeSearch, setDOM, RelativeSearchElement } = require('./proximityElementSearch');
-const devices = require('./device').default;
+const devices = require('./data/devices').default;
 const fs = require('fs');
 const os = require('os');
 const mkdtempAsync = helper.promisify(fs.mkdtemp);
